backend/pyproject.toml
[project]
name = "agent-ui-backend"
version = "0.1.0"
description = "FastAPI backend for Agent UI"
requires-python = "==3.12.*"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "passlib[argon2]>=1.7",
  "python-jose[cryptography]>=3.3",
  "pydantic>=2.7",
  "email-validator>=2.1",
  "python-dotenv>=1.0",
  "httpx>=0.27",
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.5",
  "pytest>=8.0",
]
